+++ CHANGELOG.md
@@ -13,4 +13,4 @@
 ## 1.0.5
 * dart format improvment.
 ## 1.0.6
-* adjust padding on textfield using border.
+* adjust padding on textfield using border. Add samples code